// model/collection.tsx
/**
 * Collection model
 */
export type Collection = {
    address: string
    description: string
    discord_url: string
    external_url: string
    image_url: string
    name: string
    slug: string
    twitter_username: string
    schema_name: string
    symbol: string
    banner_image_url: string | null
    instagram_username: string | null
}